Add AStar path search

+++ src/ORMMapViewer/Model/Entitites/Graph/LengthWeight.cs
@@ -4,19 +4,19 @@ namespace ORMMapViewer.Model.Entitites
 {
 	internal class LengthWeight : Weight
 	{
-		private readonly double length;
+		private readonly float length;
 
-		public LengthWeight(double length)
+		public LengthWeight(float length)
 		{
 			this.length = length;
 		}
 
 		public LengthWeight(Node first, Node second)
 		{
-			length = Math.Sqrt(Math.Pow(first.pos.X - second.pos.X, 2) + Math.Pow(first.pos.Y - second.pos.Y, 2));
+			length = (float) Math.Sqrt(Math.Pow(first.pos.X - second.pos.X, 2) + Math.Pow(first.pos.Y - second.pos.Y, 2));
 		}
 
-		public override double Calculate()
+		public override float Calculate()
 		{
 			return length;
 		}
